+++ src/app/_lib/getData.ts
@@ -1,0 +1,11 @@
+async function getData() {
+  const res = await fetch(`${process.env.NEXTAUTH_URL}/api/blog`, {
+    cache: "no-store",
+  });
+
+  if (!res.ok) {
+    throw new Error("Failed to fetch data");
+  }
+
+  return res.json();
+}
